backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pgvector
google-generativeai
python-dotenv
email-validator
PyMuPDF
alembic 